loader mixin
